+++ engine/src/flutter/lib/web_ui/lib/src/engine/navigation/history.dart
@@ -152,7 +152,7 @@ class MultiEntriesBrowserHistory extends BrowserHistory {
 
   Object _tagWithSerialCount(Object? originialState, int count) {
     return <dynamic, dynamic>{
-      'serialCount': count,
+      'serialCount': count.toDouble(),
       'state': originialState,
     };
   }
@@ -220,7 +220,7 @@ class MultiEntriesBrowserHistory extends BrowserHistory {
     assert(_hasSerialCount(currentState));
     final int backCount = _currentSerialCount;
     if (backCount > 0) {
-      await urlStrategy!.go(-backCount);
+      await urlStrategy!.go(-backCount.toDouble());
     }
     // Unwrap state.
     assert(_hasSerialCount(currentState) && _currentSerialCount == 0);
